pyproject.toml
[build-system]
requires = ["poetry==1.2.0b1"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "city-search"
version = "0.1.0"
description = "City-search API backend"
authors = ["Artyom Fomenko <artyom.fomenko@distribusion.com>"]


[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "*"
gunicorn = "*"
fastapi = "*"
fastapi-utils = "*"
typer = "*"
tortoise-orm = "*"
aerich = "*"
itsdangerous = "*"
asyncpg = "*"
python-slugify = "*"
aiohttp = "*"


[tool.poetry.scripts]
city-search = 'city_search.cli:app'


[tool.poetry.group.test]
optional = true


[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"


[tool.isort]
profile = "black"


[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
exclude = ["env/", ".venv/", "venv/"]


[tool.pylint.messages_control]
disable = "C0115, R0903, C0116, R1734, R1735, E0401"


[tool.pylint.format]
max-line-length = "88"


[tool.black]
include = "city_search"


[tool.coverage.run]
source = ["city_search"]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "strict"
